Move writer to IO namespace.
