+++ src/AsmResolver/IO/IBinaryStreamWriter.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Text;
 
-namespace AsmResolver
+namespace AsmResolver.IO
 {
     /// <summary>
     /// Provides methods for writing data to a binary stream.
@@ -119,13 +119,13 @@ namespace AsmResolver
                 writer.WriteUInt64(0);
                 count -= sizeof(ulong);
             }
-            
+
             while (count >= sizeof(uint))
             {
                 writer.WriteUInt32(0);
                 count -= sizeof(uint);
             }
-            
+
             while (count >= sizeof(ushort))
             {
                 writer.WriteUInt16(0);
@@ -181,7 +181,7 @@ namespace AsmResolver
                     throw new ArgumentOutOfRangeException(nameof(size), size, null);
             }
         }
-        
+
         /// <summary>
         /// Compresses and writes an unsigned integer to the stream.
         /// </summary>
